requirements_latex.txt
bibtexparser==1.4.2
goose3==3.1.19
langdetect==1.0.9
llama_index==0.11.23
nltk==3.9.1
openai==1.54.4
pandas==2.2.3
polyglot==16.7.4
pylatex==1.4.2
pylatexenc==2.10
python-dotenv==1.0.1
requests==2.32.3
spacy==3.8.3
pyicu==2.14
pycld2==0.41
wheel==0.45.0
setuptools==75.5.0
SQLAlchemy==2.0.36
numpy==1.23.5
